requirements.txt
absl-py==1.4.0
aiohttp @ file:///C:/b/abs_c4zmy2l696/croot/aiohttp_1670009573673/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
astor==0.8.1
async-timeout @ file:///C:/b/abs_43ozhz2a8g/croots/recipe/async-timeout_1664876362767/work
asynctest==0.13.0
attrs @ file:///C:/b/abs_09s3y775ra/croot/attrs_1668696195628/work
av==10.0.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.3.0
celluloid==0.2.0
certifi==2022.12.7
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
chardet==5.1.0
charset-normalizer==2.1.1
clang==5.0
click==8.0.3
colorama==0.4.6
conda==4.12.0
conda-package-handling @ file:///C:/b/abs_fcga8w0uem/croot/conda-package-handling_1672865024290/work
conda_package_streaming @ file:///C:/b/abs_0e5n5hdal3/croot/conda-package-streaming_1670508162902/work
cryptography @ file:///C:/b/abs_b7d7drzbky/croot/cryptography_1673298763653/work
cycler==0.10.0
deepdish==0.3.7
dlib @ file:///D:/bld/dlib_1652035402968/work
docopt==0.6.2
easing-functions==1.0.4
flatbuffers==23.1.21
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
frozenlist @ file:///C:/b/abs_2bb5uzghsi/croot/frozenlist_1670004511812/work
gast==0.5.3
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///C:/ci/grpcio_1637590993074/work
h5py==2.8.0
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
imageio==2.26.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1648562631189/work
imutils==0.5.2
joblib==1.0.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kornia==0.6.10
libclang==15.0.6.1
lxml==4.9.2
Mako @ file:///C:/b/abs_e2_71_9f9g/croot/mako_1665472436410/work
mamba==0.1.2
Markdown @ file:///C:/b/abs_98lv_ucina/croot/markdown_1671541919225/work
MarkupSafe @ file:///C:/ci/markupsafe_1654508076077/work
matplotlib==3.3.0
menuinst @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_455sf5o0ct/croots/recipe/menuinst_1661805970842/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
mock==4.0.3
multidict @ file:///C:/b/abs_6cx_8w3cv2/croot/multidict_1665674238352/work
natsort==8.3.1
networkx==2.6.3
nibabel==4.0.2
nilearn==0.10.0
nltools==0.4.7
numexpr==2.8.4
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.1.0.25
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
pipreqs==0.4.11
protobuf==3.20.3
py-feat==0.5.1
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycosat @ file:///C:/b/abs_4b1rrw8pn9/croot/pycosat_1666807711599/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydot==1.4.2
pygpu==0.7.6
PyJWT @ file:///C:/ci/pyjwt_1657529476747/work
pynv==0.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==2.4.7
PyQt5==5.11.3
pyqt5-tools==5.11.3.1.4
PyQt5_sip==4.19.19
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
PyWavelets==1.3.0
PyYAML==5.4.1
requests @ file:///C:/ci/requests_1657735288441/work
requests-oauthlib==1.3.1
rsa==4.9
ruamel_yaml==0.15.87
scikit-image==0.19.3
scikit-learn==1.0.2
scipy @ file:///C:/ci/scipy_1661333074914/work
seaborn==0.12.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
tables==3.7.0
tensorboard @ file:///C:/Users/builder/adipietro/mc3/tf210/conda-bld/tensorboard_1669760968711/work/tensorboard-2.10.0-py3-none-any.whl
tensorboard-data-server @ file:///C:/b/abs_2fhvpo862s/croot/tensorboard-data-server_1670853600144/work/tensorboard_data_server-0.6.1-py3-none-any.whl
tensorboard-plugin-wit @ file:///C:/tf/b/tensorboard-plugin-wit_1660162132996/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-io-gcs-filesystem==0.30.0
termcolor==1.1.0
Theano==1.0.5
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions @ file:///C:/b/abs_89eui86zuq/croot/typing_extensions_1669923792806/work
urllib3 @ file:///C:/b/abs_9bcwxczrvm/croot/urllib3_1673575521331/work
Werkzeug @ file:///C:/b/abs_17q5kgb8bo/croot/werkzeug_1671216014857/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.12.1
xgboost==1.6.2
yarg==0.1.9
yarl @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_e5nlunspj6/croots/recipe/yarl_1661437086516/work
zipp @ file:///C:/b/abs_b9jfdr908q/croot/zipp_1672387552360/work
zstandard==0.18.0
